+++ test/spec/selfie/manage_campaign_placement_tag.controller.ut.js
@@ -14,6 +14,7 @@ define(['app','c6_defines'], function(appModule, c6Defines) {
             module(appModule.name);
 
             c6Defines.kPlatformHome = 'https://platform-staging.reelcontent.com/';
+            c6Defines.kAuditUrl = 'https://audit-staging.reelcontent.com/';
 
             inject(function($injector) {
                 $rootScope = $injector.get('$rootScope');
@@ -57,6 +58,7 @@ define(['app','c6_defines'], function(appModule, c6Defines) {
 
                         expect(SelfieManageCampaignPlacementTagCtrl.placement).toEqual({
                             name: 'beeswax',
+                            pixel: jasmine.any(String),
                             tag: [
                                 '<div>',
                                 '    <script>',
@@ -101,6 +103,7 @@ define(['app','c6_defines'], function(appModule, c6Defines) {
 
                         expect(SelfieManageCampaignPlacementTagCtrl.placement).toEqual({
                             name: 'beeswax',
+                            pixel: jasmine.any(String),
                             tag: [
                                 '<div>',
                                 '    <script>',
@@ -123,6 +126,69 @@ define(['app','c6_defines'], function(appModule, c6Defines) {
                             ].join('\n')
                         });
                     });
+
+                    it('should generate the impression pixel', function() {
+                        SelfieManageCampaignPlacementTagCtrl.initWithModel({
+                            id: 'pl-111',
+                            tagType: 'mraid',
+                            tagParams: {
+                                container: 'beeswax',
+                                campaign: 'cam-123'
+                            },
+                            showInTag: {}
+                        });
+
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain(c6Defines.kAuditUrl + 'pixel.gif?');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('cb=1');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&placement=pl-111');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&event=impression');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&container=beeswax');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&campaign=cam-123');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&hostApp=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&network=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&branding=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&ex=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&vr=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&uuid=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&domain=');
+
+                        SelfieManageCampaignPlacementTagCtrl.initWithModel({
+                            id: 'pl-111',
+                            tagType: 'mraid',
+                            tagParams: {
+                                container: 'beeswax',
+                                campaign: 'cam-123',
+                                clickUrls: ['${click}', '{{CLICK}}'],
+                                countdown: 30,
+                                branding: 'mybrand',
+                                prebuffer: true,
+                                forceOrientation: 'portrait',
+                                network: '${network}',
+                                uuid: 'user',
+                                ex: 'something',
+                                vr: 'else',
+                                hostApp: '{{HOST_APP}}',
+                                domain: '{{page_url}}'
+                            },
+                            showInTag: {}
+                        });
+
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain(c6Defines.kAuditUrl + 'pixel.gif?');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('cb=1');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&placement=pl-111');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&event=impression');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&container=beeswax');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&hostApp={{HOST_APP}}');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&network=${network}');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&branding=mybrand');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&ex=something');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&vr=else');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&externalSessionId=user');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&domain={{page_url}}');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&countdown=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&prebuffer=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&forceOrientation=');
+                    });
                 });
 
                 describe('when placement type is VPAID', function() {
@@ -138,6 +204,7 @@ define(['app','c6_defines'], function(appModule, c6Defines) {
 
                         expect(SelfieManageCampaignPlacementTagCtrl.placement).toEqual({
                             name: 'beeswax',
+                            pixel: jasmine.any(String),
                             tag: c6Defines.kPlatformHome + 'api/public/vast/2.0/tag?placement=pl-111'
                         });
                     });
@@ -167,8 +234,72 @@ define(['app','c6_defines'], function(appModule, c6Defines) {
 
                         expect(SelfieManageCampaignPlacementTagCtrl.placement).toEqual({
                             name: 'beeswax',
+                            pixel: jasmine.any(String),
                             tag: c6Defines.kPlatformHome + 'api/public/vast/2.0/tag?placement=pl-111&clickUrls=${click},{{CLICK}}&countdown=30&prebuffer=true&network=${network}'
                         });
+                    });
+
+                    it('should generate the impression pixel', function() {
+                        SelfieManageCampaignPlacementTagCtrl.initWithModel({
+                            id: 'pl-111',
+                            tagType: 'vpaid',
+                            tagParams: {
+                                container: 'beeswax',
+                                campaign: 'cam-123'
+                            },
+                            showInTag: {}
+                        });
+
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain(c6Defines.kAuditUrl + 'pixel.gif?');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('cb=1');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&placement=pl-111');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&event=dspimpression');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&container=beeswax');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&campaign=cam-123');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&hostApp=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&network=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&branding=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&ex=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&vr=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&uuid=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&domain=');
+
+                        SelfieManageCampaignPlacementTagCtrl.initWithModel({
+                            id: 'pl-111',
+                            tagType: 'vpaid',
+                            tagParams: {
+                                container: 'beeswax',
+                                campaign: 'cam-123',
+                                clickUrls: ['${click}', '{{CLICK}}'],
+                                countdown: 30,
+                                branding: 'mybrand',
+                                prebuffer: true,
+                                forceOrientation: 'portrait',
+                                network: '${network}',
+                                uuid: 'user',
+                                ex: 'something',
+                                vr: 'else',
+                                hostApp: '{{HOST_APP}}',
+                                domain: '{{page_url}}'
+                            },
+                            showInTag: {}
+                        });
+
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain(c6Defines.kAuditUrl + 'pixel.gif?');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('cb=1');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&placement=pl-111');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&event=dspimpression');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&container=beeswax');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&hostApp={{HOST_APP}}');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&network=${network}');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&branding=mybrand');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&ex=something');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&vr=else');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&externalSessionId=user');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).toContain('&domain={{page_url}}');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&countdown=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&prebuffer=');
+                        expect(SelfieManageCampaignPlacementTagCtrl.placement.pixel).not.toContain('&forceOrientation=');
                     });
                 });
             });